requirements.txt
groq
python-dotenv
openai